requirements.txt
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
groq>=0.9.0
openai>=1.30.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
rich>=13.0.0
httpx>=0.27.0
